// v2/public.d.ts
export interface PublicKey extends Key {
    readonly algorithm: {
        readonly name: 'PASETO v2.public';
    };
    readonly type: 'public';
    readonly version: 2;
    readonly kind: 'public';
}
// ...
export interface SecretKey extends Key {
    readonly algorithm: {
        readonly name: 'PASETO v2.public';
    };
    readonly type: 'secret';
    readonly version: 2;
    readonly kind: 'secret';
}
// ...
export declare const SignFactory: ReturnType<typeof PublicSign<2, SecretKey>>;
/**
 * Built-in v2.public token-verification capability factory.
 *
 * @category Token Operations
 */
export declare const VerifyFactory: ReturnType<typeof PublicVerify<2, PublicKey>>;
// ...
export declare const ExportSecretKeyFactory: ReturnType<typeof PublicExportSecretKey<2, SecretKey>>;
/**
 * Built-in PASERK k2 public-key derivation capability factory.
 *
 * @category Key Management
 */
export declare const GetPublicKeyFactory: ReturnType<typeof PublicGetPublicKey<2, PublicKey, SecretKey>>;